created styles folder and moved globals.css
