--- app/components/Footer.js
@@ -1,3 +1,0 @@
-export default function Footer() {
-  return <footer>This is my footer</footer>
-}
