+++ app/layout.js
@@ -1,6 +1,6 @@
-import "./globals.css";
+import "../styles/globals.css";
 import { Inter } from "next/font/google";
-import Navbar from "./components/Navbar";
+import Navbar from "../components/Navbar";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -14,10 +14,10 @@ export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body className={inter.className}>
-        <div className="flex min-h-screen py-10">
+        {/* <div className="flex min-h-screen py-10"> */}
           <Navbar />
           {children}
-        </div>
+        {/* </div> */}
       </body>
     </html>
   );
